+++ HISTORY.md
@@ -3,6 +3,7 @@
 
 ## not yet released, version 2.0.0-SNAPSHOT
 
+- Implemented support for fractions, powered by the library `fraction.js`.
 - Implemented matrix LU decomposition with partial pivoting and a LU based 
   linear equations solver (functions `lup` and `lusolve`). Thanks @rjbaucells.
 - Large internal refactoring, allowing to create custom bundles of math.js.